requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.1
Pillow==10.3.0
flask==3.0.0
flask-cors==4.0.0